Fix the ci failed test cases for the schema evolution
